Setup the romhack injection

+++ dw/build.sh
@@ -10,6 +10,13 @@ magenta=$(tput setaf 5)
 cyan=$(tput setaf 6)
 reset=$(tput sgr0)
 
+# Check for required subfolders
+if [ ! -d "asm/bin" ]; then
+    echo "ERROR: asm/bin/ folder not found."
+    echo "You should run extract.sh before running the build."
+    exit 1
+fi
+
 printf "\n${magenta}Deleting previous build files...${reset}\n"
 rm -rf build
 mkdir -p build
@@ -18,18 +25,17 @@ printf "\n${magenta}Assembling with ca65...${reset}\n\n"
 
 # Header (should emit exactly 16 bytes from $0000)
 ca65 -o build/Header.o asm/Header.asm
-ld65 -C asm/header.cfg -o build/Header.bin build/Header.o
 
 # PRG banks
 ca65 -o build/Bank00.o asm/Bank00.asm
-ca65 -o build/Bank01.o asm/Bank01.asm
+ca65 -o build/Bank01.o asm/Bank01.asm -D retroai
 ca65 -o build/Bank02.o asm/Bank02.asm
 ca65 -o build/Bank03.o asm/Bank03.asm
 
-ld65 -C asm/prg_bank.cfg -o build/Bank00.bin build/Bank00.o
-ld65 -C asm/prg_bank.cfg -o build/Bank01.bin build/Bank01.o
-ld65 -C asm/prg_bank.cfg -o build/Bank02.bin build/Bank02.o
-ld65 -C asm/prg_bank.cfg -o build/Bank03.bin build/Bank03.o
+ld65 -C asm/nes.cfg \
+    build/Header.o \
+    build/Bank00.o build/Bank01.o build/Bank02.o build/Bank03.o \
+    --mapfile build/map.txt
 
 # ------------------------------------------------------------------------------
 # Combine into final .nes:
@@ -47,8 +53,8 @@ cat build/Header.bin \
     > build/DragonWarrior.nes
 
 printf "${magenta}Cleaning up intermediate files......${reset}\n"	
-#rm -f build/*.o
-#rm -f build/*.bin
+rm -f build/*.o
+rm -f build/*.bin
 
 printf "\n${magenta}Verifying final ROM checksum...${reset}\n\n"
 final_md5=($(md5sum build/DragonWarrior.nes))
@@ -60,6 +66,9 @@ if [ "$final_md5" = "$expected_md5" ]; then
 else
     printf "${red}Final ROM checksum mismatch!${reset}\n\n"
 fi
+
+printf "\n${magenta}Running size check...${reset}\n\n"
+bash ./scripts/check_limits.sh
 
 printf "${green}Build complete.${reset}\n"
 printf "Output written to: build/DragonWarrior.nes\n"
